:hammer: split "parsing" directory to "input" and "output"
