+++ develop/detail/output/output_adapters.hpp
