+++ src/json.hpp
@@ -1543,7 +1543,7 @@ constexpr const auto& to_json = detail::static_const<detail::to_json_fn>::value;
 }
 }
 
-// #include "detail/parsing/input_adapters.hpp"
+// #include "detail/input/input_adapters.hpp"
 
 
 #include <algorithm> // min
@@ -1808,7 +1808,7 @@ class input_adapter
 }
 }
 
-// #include "detail/parsing/lexer.hpp"
+// #include "detail/input/lexer.hpp"
 
 
 #include <clocale> // localeconv
@@ -1823,7 +1823,7 @@ class input_adapter
 
 // #include "detail/macro_scope.hpp"
 
-// #include "detail/parsing/input_adapters.hpp"
+// #include "detail/input/input_adapters.hpp"
 
 
 namespace nlohmann
@@ -3088,7 +3088,7 @@ scan_number_done:
 }
 }
 
-// #include "detail/parsing/parser.hpp"
+// #include "detail/input/parser.hpp"
 
 
 #include <cassert> // assert
@@ -3102,9 +3102,9 @@ scan_number_done:
 
 // #include "detail/macro_scope.hpp"
 
-// #include "detail/parsing/input_adapters.hpp"
+// #include "detail/input/input_adapters.hpp"
 
-// #include "detail/parsing/lexer.hpp"
+// #include "detail/input/lexer.hpp"
 
 // #include "detail/value_t.hpp"
 
@@ -4678,7 +4678,7 @@ class json_reverse_iterator : public std::reverse_iterator<Base>
 }
 }
 
-// #include "detail/parsing/output_adapters.hpp"
+// #include "detail/output/output_adapters.hpp"
 
 
 #include <algorithm> // copy
@@ -4793,7 +4793,7 @@ class output_adapter
 }
 }
 
-// #include "detail/parsing/binary_reader.hpp"
+// #include "detail/input/binary_reader.hpp"
 
 
 #include <algorithm> // generate_n
@@ -4814,8 +4814,6 @@ class output_adapter
 // #include "detail/exceptions.hpp"
 
 // #include "detail/macro_scope.hpp"
-
-// #include "detail/parsing/input_adapters.hpp"
 
 // #include "detail/value_t.hpp"
 
@@ -4864,7 +4862,7 @@ class binary_reader
         if (strict)
         {
             get();
-            check_eof(true);
+            expect_eof();
         }
         return res;
     }
@@ -4885,7 +4883,7 @@ class binary_reader
         if (strict)
         {
             get();
-            check_eof(true);
+            expect_eof();
         }
         return res;
     }
@@ -4906,7 +4904,7 @@ class binary_reader
         if (strict)
         {
             get_ignore_noop();
-            check_eof(true);
+            expect_eof();
         }
         return res;
     }
@@ -5005,7 +5003,6 @@ class binary_reader
 
             case 0x38: // Negative integer (one-byte uint8_t follows)
             {
-                // must be uint8_t !
                 return static_cast<number_integer_t>(-1) - get_number<uint8_t>();
             }
 
@@ -5196,9 +5193,9 @@ class binary_reader
             case 0xF9: // Half-Precision Float (two-byte IEEE 754)
             {
                 const int byte1 = get();
-                check_eof();
+                unexpect_eof();
                 const int byte2 = get();
-                check_eof();
+                unexpect_eof();
 
                 // code from RFC 7049, Appendix D, Figure 3:
                 // As half-precision floating-point numbers were only added
@@ -5631,7 +5628,7 @@ class binary_reader
         for (std::size_t i = 0; i < sizeof(NumberType); ++i)
         {
             get();
-            check_eof();
+            unexpect_eof();
 
             // reverse byte order prior to conversion if necessary
             if (is_little_endian)
@@ -5656,7 +5653,7 @@ class binary_reader
     @param[in] len number of bytes to read
 
     @note We can not reserve @a len bytes for the result, because @a len
-          may be too large. Usually, @ref check_eof() detects the end of
+          may be too large. Usually, @ref unexpect_eof() detects the end of
           the input before we run out of string memory.
 
     @return string created by reading @a len bytes
@@ -5670,7 +5667,7 @@ class binary_reader
         std::generate_n(std::back_inserter(result), len, [this]()
         {
             get();
-            check_eof();
+            unexpect_eof();
             return static_cast<char>(current);
         });
         return result;
@@ -5690,7 +5687,7 @@ class binary_reader
     */
     std::string get_cbor_string()
     {
-        check_eof();
+        unexpect_eof();
 
         switch (current)
         {
@@ -5748,7 +5745,7 @@ class binary_reader
                 std::string result;
                 while (get() != 0xFF)
                 {
-                    check_eof();
+                    unexpect_eof();
                     result.push_back(static_cast<char>(current));
                 }
                 return result;
@@ -5803,7 +5800,7 @@ class binary_reader
     */
     std::string get_msgpack_string()
     {
-        check_eof();
+        unexpect_eof();
 
         switch (current)
         {
@@ -5919,7 +5916,7 @@ class binary_reader
             get();  // TODO: may we ignore N here?
         }
 
-        check_eof();
+        unexpect_eof();
 
         switch (current)
         {
@@ -5959,7 +5956,7 @@ class binary_reader
         if (current == '$')
         {
             tc = get();  // must not ignore 'N', because 'N' maybe the type
-            check_eof();
+            unexpect_eof();
 
             get_ignore_noop();
             if (current != '#')
@@ -6012,7 +6009,7 @@ class binary_reader
             case 'C':  // char
             {
                 get();
-                check_eof();
+                unexpect_eof();
                 if (JSON_UNLIKELY(current > 127))
                 {
                     std::stringstream ss;
@@ -6121,24 +6118,26 @@ class binary_reader
     }
 
     /*!
-    @brief check if input ended
+    @brief throw if end of input is not reached
+    @throw parse_error.110 if input not ended
+    */
+    void expect_eof() const
+    {
+        if (JSON_UNLIKELY(current != std::char_traits<char>::eof()))
+        {
+            JSON_THROW(parse_error::create(110, chars_read, "expected end of input"));
+        }
+    }
+
+    /*!
+    @briefthrow if end of input is reached
     @throw parse_error.110 if input ended
     */
-    void check_eof(const bool expect_eof = false) const
+    void unexpect_eof() const
     {
-        if (expect_eof)
+        if (JSON_UNLIKELY(current == std::char_traits<char>::eof()))
         {
-            if (JSON_UNLIKELY(current != std::char_traits<char>::eof()))
-            {
-                JSON_THROW(parse_error::create(110, chars_read, "expected end of input"));
-            }
-        }
-        else
-        {
-            if (JSON_UNLIKELY(current == std::char_traits<char>::eof()))
-            {
-                JSON_THROW(parse_error::create(110, chars_read, "unexpected end of input"));
-            }
+            JSON_THROW(parse_error::create(110, chars_read, "unexpected end of input"));
         }
     }
 
@@ -6158,7 +6157,7 @@ class binary_reader
 }
 }
 
-// #include "detail/parsing/binary_writer.hpp"
+// #include "detail/output/binary_writer.hpp"
 
 
 #include <algorithm> // reverse
@@ -6167,9 +6166,9 @@ class binary_reader
 #include <cstring> // memcpy
 #include <limits> // numeric_limits
 
-// #include "detail/parsing/binary_reader.hpp"
+// #include "detail/input/binary_reader.hpp"
 
-// #include "detail/parsing/output_adapters.hpp"
+// #include "detail/output/output_adapters.hpp"
 
 
 namespace nlohmann
@@ -7070,7 +7069,7 @@ class binary_writer
 }
 }
 
-// #include "detail/serializer.hpp"
+// #include "detail/output/serializer.hpp"
 
 
 #include <algorithm> // reverse, remove, fill, find, none_of
@@ -8186,7 +8185,7 @@ char* to_chars(char* first, char* last, FloatType value)
 
 // #include "detail/meta.hpp"
 
-// #include "detail/parsing/output_adapters.hpp"
+// #include "detail/output/output_adapters.hpp"
 
 // #include "detail/value_t.hpp"
 
@@ -8993,7 +8992,7 @@ class json_pointer
     */
     static int array_index(const std::string& s)
     {
-        size_t processed_chars = 0;
+        std::size_t processed_chars = 0;
         const int res = std::stoi(s, &processed_chars);
 
         // check if the string was completely read
